fix: old debug code preventing timeouts from running properly

+++ src/widgets/tasks.cpp
@@ -43,7 +43,7 @@ void keyboard_onenter(kb_data_t* data) {
 	fflush(stdout);
 	
 	int actual_buf_len = 0;
-	for (int i=0; i < strlen(data->kb_buf) - 1; i++) {
+	for (size_t i=0; i < strlen(data->kb_buf) - 1; i++) {
 	    if (data->kb_buf[i] >= 0x2e && data->kb_buf[i] <= 0x7e) { actual_buf_len++; }
 	};
